+++ apitest-commons/src/main/java/io/mosip/testrig/apirig/utils/TestCaseReader.java
@@ -23,65 +23,65 @@ public class TestCaseReader extends BaseTestCase {
 	 * @throws ParseException
 	 */
 	
-	CommonLibrary commonLib = new CommonLibrary();
-	public Object[][] readTestCases(String folderName, String testType){
-
-		List<String> listOfFolders = commonLib.getFoldersFilesNameList(folderName, true);
-		
-		ArrayList<String> testCaseNames = new ArrayList<>();
-		
-		for (int j = 0; j < listOfFolders.size(); j++) {
-			String[] arr = listOfFolders.get(j).split("/");
-			switch (testType) {
-			case "smoke":
-				if (arr[arr.length - 1].contains("smoke"))
-					testCaseNames.add(arr[arr.length - 1]);
-				break;
-			case "regression":
-				if (arr[arr.length - 1].contains("invalid"))
-					testCaseNames.add(arr[arr.length - 1]);
-				break;
-
-			default:
-				testCaseNames.add(arr[arr.length - 1]);
-				break;
-			}
-		}
-
-		Object[][] testParam = new Object[testCaseNames.size()][];
-		int k = 0;
-		for (String testcaseName : testCaseNames) {
-			
-			testParam[k] = new Object[] {testcaseName};
-			k++;
-		}
-		return testParam;
-
-	}
-
-
-	/**
-	 * @param modulename
-	 * @param apiname
-	 * @param testcaseName
-	 * @return this method is for reading request and response object form the given testcase folder and returns in array.
-	 */
-	public JSONObject[] readRequestResponseJson(String modulename, String apiname, String testcaseName){
-		String configPath = modulename + "/" + apiname + "/" + testcaseName;
-		List<String> listofFiles =  commonLib.getFoldersFilesNameList(configPath, false);
-		JSONObject[] objectData = new JSONObject[2];
-		for (int k = 0; k < listofFiles.size(); k++) {
-			String[] arr = listofFiles.get(k).split("/");
-				if (arr[arr.length - 1].toLowerCase().contains(GlobalConstants.REQUEST)) 
-					objectData[0] =  commonLib.readJsonData(configPath+"/"+arr[arr.length - 1], true);
-					
-				 else if (arr[arr.length - 1].toLowerCase().contains("response")) 
-					objectData[1] =  commonLib.readJsonData(configPath+"/"+arr[arr.length - 1], true);
-		 
-			
-		}
-		return objectData;
-	}
+//	CommonLibrary commonLib = new CommonLibrary();
+//	public Object[][] readTestCases(String folderName, String testType){
+//
+//		List<String> listOfFolders = commonLib.getFoldersFilesNameList(folderName, true);
+//		
+//		ArrayList<String> testCaseNames = new ArrayList<>();
+//		
+//		for (int j = 0; j < listOfFolders.size(); j++) {
+//			String[] arr = listOfFolders.get(j).split("/");
+//			switch (testType) {
+//			case "smoke":
+//				if (arr[arr.length - 1].contains("smoke"))
+//					testCaseNames.add(arr[arr.length - 1]);
+//				break;
+//			case "regression":
+//				if (arr[arr.length - 1].contains("invalid"))
+//					testCaseNames.add(arr[arr.length - 1]);
+//				break;
+//
+//			default:
+//				testCaseNames.add(arr[arr.length - 1]);
+//				break;
+//			}
+//		}
+//
+//		Object[][] testParam = new Object[testCaseNames.size()][];
+//		int k = 0;
+//		for (String testcaseName : testCaseNames) {
+//			
+//			testParam[k] = new Object[] {testcaseName};
+//			k++;
+//		}
+//		return testParam;
+//
+//	}
+//
+//
+//	/**
+//	 * @param modulename
+//	 * @param apiname
+//	 * @param testcaseName
+//	 * @return this method is for reading request and response object form the given testcase folder and returns in array.
+//	 */
+//	public JSONObject[] readRequestResponseJson(String modulename, String apiname, String testcaseName){
+//		String configPath = modulename + "/" + apiname + "/" + testcaseName;
+//		List<String> listofFiles =  commonLib.getFoldersFilesNameList(configPath, false);
+//		JSONObject[] objectData = new JSONObject[2];
+//		for (int k = 0; k < listofFiles.size(); k++) {
+//			String[] arr = listofFiles.get(k).split("/");
+//				if (arr[arr.length - 1].toLowerCase().contains(GlobalConstants.REQUEST)) 
+//					objectData[0] =  commonLib.readJsonData(configPath+"/"+arr[arr.length - 1], true);
+//					
+//				 else if (arr[arr.length - 1].toLowerCase().contains("response")) 
+//					objectData[1] =  commonLib.readJsonData(configPath+"/"+arr[arr.length - 1], true);
+//		 
+//			
+//		}
+//		return objectData;
+//	}
 	
 	
 }
